Moved PropertyReplacerStrategy to right package
